Integration tests moved to internal Python repo
