+++ pkg/recorder/recorder_test.go
@@ -1,15 +1,42 @@
 package recorder
 
 import (
+	"context"
+	"encoding/json"
 	"fmt"
 	"testing"
 	"time"
 
 	"github.com/openshift/insights-operator/pkg/record"
-	"github.com/openshift/insights-operator/tests"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
+
+// RawReport implements Marshable interface
+type RawReport struct{ Data string }
+
+// Marshal returns raw bytes
+func (r RawReport) Marshal(_ context.Context) ([]byte, error) {
+	return []byte(r.Data), nil
+}
+
+// GetExtension returns extension for raw marshaller
+func (r RawReport) GetExtension() string {
+	return ""
+}
+
+// RawInvalidReport implements Marshable interface but throws an error
+type RawInvalidReport struct{}
+
+// Marshal returns raw bytes
+func (r RawInvalidReport) Marshal(_ context.Context) ([]byte, error) {
+	return nil, &json.UnsupportedTypeError{}
+}
+
+// GetExtension returns extension for raw marshaller
+func (r RawInvalidReport) GetExtension() string {
+	return ""
+}
 
 type driverMock struct {
 	mock.Mock
@@ -44,7 +71,7 @@ func Test_Record(t *testing.T) {
 	rec := newRecorder()
 	err := rec.Record(record.Record{
 		Name: "config/mock1",
-		Item: tests.RawReport{Data: "mock1"},
+		Item: RawReport{Data: "mock1"},
 	})
 	assert.Nil(t, err)
 	assert.Equal(t, 1, len(rec.records))
@@ -54,12 +81,12 @@ func Test_Record_Duplicated(t *testing.T) {
 	rec := newRecorder()
 	_ = rec.Record(record.Record{
 		Name:        "config/mock1",
-		Item:        tests.RawReport{Data: "mock1"},
+		Item:        RawReport{Data: "mock1"},
 		Fingerprint: "abc",
 	})
 	err := rec.Record(record.Record{
 		Name:        "config/mock1",
-		Item:        tests.RawReport{Data: "mock1"},
+		Item:        RawReport{Data: "mock1"},
 		Fingerprint: "abc",
 	})
 	assert.Nil(t, err)
@@ -70,7 +97,7 @@ func Test_Record_CantBeSerialized(t *testing.T) {
 	rec := newRecorder()
 	err := rec.Record(record.Record{
 		Name: "config/mock1",
-		Item: tests.RawInvalidReport{},
+		Item: RawInvalidReport{},
 	})
 	assert.Error(t, err)
 }
@@ -80,7 +107,7 @@ func Test_Record_Flush(t *testing.T) {
 	for i := range []int{1, 2, 3} {
 		_ = rec.Record(record.Record{
 			Name: fmt.Sprintf("config/mock%d", i),
-			Item: tests.RawReport{Data: "mockdata"},
+			Item: RawReport{Data: "mockdata"},
 		})
 	}
 	err := rec.Flush()
