--- tests/integration/resource_samples/apply.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-# should be run from project root directory
-for file in ./tests/integration/resource_samples/*.yaml
-do
-  oc apply -f $file
-done
