--- tests/integration/resource_samples/csr.yaml
@@ -1,13 +1,0 @@
-apiVersion: certificates.k8s.io/v1
-kind: CertificateSigningRequest
-metadata:
-  name: service.default
-spec:
-  signerName: www.example.com/example
-  groups:
-    - system:authenticated
-  request: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURSBSRVFVRVNULS0tLS0KTUlJQ2xqQ0NBWDRDQVFBd1VURUxNQWtHQTFVRUJoTUNZM294RFRBTEJnTlZCQWdNQkhCcFkya3hGVEFUQmdOVgpCQWNNREVSbFptRjFiSFFnUTJsMGVURWNNQm9HQTFVRUNnd1RSR1ZtWVhWc2RDQkRiMjF3WVc1NUlFeDBaRENDCkFTSXdEUVlKS29aSWh2Y05BUUVCQlFBRGdnRVBBRENDQVFvQ2dnRUJBTHdSTDFoT3lOTERONi9CVkZPeldoSXoKZmhmN2F1a0xPL2JWa0FVb2JPcjlwT0VJbURZemV3MFpwUkhnd2t5eHFDVVUrYVY1M2p1Q3RBSm8zZCtPVTBDbQpoVmZScFM3Z09DMUdvdFN6ekh1RU5YU2l1bzczWGxCQXhEOUpMWW1xUDdXNTNYK3pLaXN3L2crUGxxYTV1LzlECi9qWUJ3QXEzSm5zT2VMckZ0cVVrOThTaTVyUTQ3WVhNQ0N2UFRhT05US08wMlB4Z3MwbEgwVEVUOTMxOEc1VFIKM25rNy9uUWR4eERCc0FxK2UzRVU1cjEzYUw4OWN1T3dEZFJGblpaTFBzcmFPNEVqeTAzWTRVRWNVUFVxSlA4MQpQT0FsWG9PT1pYcFJIZmpBNnlLbE1aNWsxd3FHU0xXY21aYTVTTTFvamZaK21USVhyYkNDeGZneHYzazc1WmtDCkF3RUFBYUFBTUEwR0NTcUdTSWIzRFFFQkN3VUFBNElCQVFCOHp5M25hY2grMG50NGxsNUw3cnpyTnZlOWhKVFgKa1gyQXpNOU82WVZSRnJMalZsb25xU01XT2RLLzF0RU1QdFB6T1dhazN5OEEzeHE5RUN2OWtwV0ZVVnpZUFROVApzYXF5ai9MU1VpM1RmQWZsRitlUXl1VE0wZldCbEpQeS9Ya1dUeWxXbGRFNkd3TjdSSGdWYytEaC8rdjR1eDFpCklyd2ZlRjI4cHY1V3M4NHZzR0N1Mmh6SzMyNkpHSkcvMDdNUmVLL0lnclEwS21GM3h2cGJJdlFMOUl1Nldydy8KQVp5VFNCcnErblA1MjlKeWw2ZFdKaTRldDZSY2ovRzVxV2NpL1F1YzBsSUFoNGE3NVpNeCt2eGhsMUdRL0h5RQo2UE9YVCtPeGRGOTZzZnJMTnNLMzlMWElkdmRkdngxTTVFS1A3VHMrSFdDcnJoMThBKzFDaUdpYgotLS0tLUVORCBDRVJUSUZJQ0FURSBSRVFVRVNULS0tLS0K
-  usages:
-    - digital signature
-    - key encipherment
-    - server auth
